Move the bound-uuid-list idiom into a uuid_list SQL macro so the delete statements are fully static text with no Go concatenation.
