+++ desktopexporter/internal/store/util/util.go
@@ -43,43 +43,6 @@ func BuildPlaceholders(count int) string {
 	return buildPlaceholders(count, "?")
 }
 
-// UUIDList renders a set of uuid strings as one bound-list predicate body.
-//
-// The SQL it goes into is static: `where id in (` + UUIDList() + `)`, with the
-// whole set travelling as a single argument. That removes the class of bug the
-// previous form invited, where the number of `?::uuid` placeholders built into
-// the text and the number of arguments appended beside it could drift apart --
-// a mismatch SQL cannot catch, because both counts are correct on their own.
-//
-// The list binds as varchar[] and is cast inside the query rather than bound as
-// uuid[]. That is not a stylistic choice: binding the driver's duckdb.UUID type
-// as a *parameter* silently returns nothing. Measured -- a bound
-// 11111111-1111-1111-1111-111111111111 reads back as 91111111-..., matching no
-// row and raising no error, for both a scalar parameter and a list. It is safe
-// through the appender, which is the only place this codebase uses the type,
-// and every query path already binds uuids as strings.
-//
-// The likely cause, offered as a lead rather than a fact: DuckDB documents
-// UUIDs as "represented internally as HUGEINT values", which is signed, so
-// something has to flip the top bit to keep unsigned UUID ordering intact --
-// and the corruption observed is exactly bit 127. That flip is not documented
-// and is not visible from SQL (UUIDs sort 0, 1, 9, f as you would expect, and
-// the UUID -> HUGEINT cast is not implemented), so treat the mechanism as
-// unconfirmed. The behaviour is not: do not bind duckdb.UUID as a parameter.
-//
-// Watch this. The double cast is a workaround for a driver limitation, not a
-// property of the schema, so it should not outlive the limitation.
-// TestDriverStillCannotBindUUIDLists pins both halves and fails the moment
-// either lifts -- at which point this collapses to `select unnest(?::uuid[])`
-// and the comment above can go with it.
-//
-// Measured at 204,891 spans, deletes rolled back between rounds: within noise
-// of the placeholder form at every size (0.3ms vs 0.5ms at one id, 22.4ms vs
-// 21.8ms at a thousand), so the safety costs nothing worth counting.
-func UUIDList() string {
-	return "select unnest(?::varchar[])::uuid"
-}
-
 func buildPlaceholders(count int, mark string) string {
 	if count == 0 {
 		return ""
