+++ desktopexporter/internal/store/util/util_test.go
@@ -1,12 +1,9 @@
 package util
 
 import (
-	"database/sql"
 	"testing"
 
-	_ "github.com/duckdb/duckdb-go/v2"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestCamelToSnake(t *testing.T) {
@@ -28,38 +25,5 @@ func TestCamelToSnake(t *testing.T) {
 			got := CamelToSnake(tt.in)
 			assert.Equal(t, tt.want, got)
 		})
-	}
-}
-
-// The uuid placeholder contract, pinned against a real database because the
-// interesting behaviour is DuckDB's, not ours: both id spellings the API serves
-// must delete, and a malformed id must raise rather than report a successful
-// no-op.
-func TestUUIDListRejectsMalformedIDs(t *testing.T) {
-	db, err := sql.Open("duckdb", "")
-	require.NoError(t, err)
-	defer db.Close()
-	_, err = db.Exec(`create table t (id uuid)`)
-	require.NoError(t, err)
-	_, err = db.Exec(`insert into t values ('720cea1f-6c57-2438-d9b1-098fa86ecc3b')`)
-	require.NoError(t, err)
-
-	query := `select count(*) from t where id in (` + UUIDList() + `)`
-
-	for _, valid := range []string{
-		"720cea1f-6c57-2438-d9b1-098fa86ecc3b", // dashed
-		"720cea1f6c572438d9b1098fa86ecc3b",     // the wire form the API serves
-	} {
-		var n int
-		require.NoError(t, db.QueryRow(query, []string{valid}).Scan(&n), "id %q", valid)
-		assert.Equal(t, 1, n, "id %q must match the stored row", valid)
-	}
-
-	for _, bad := range []string{"not-a-uuid", ""} {
-		var n int
-		err := db.QueryRow(query, []string{bad}).Scan(&n)
-		assert.Error(t, err,
-			"a malformed id (%q) must raise, not silently match nothing -- a delete "+
-				"reporting success having removed nothing is worse than an error", bad)
 	}
 }
